Move tests to integration test crate
